requirements.txt
aiofiles>=24.1.0,<25
aiohttp>=3.12.11,<4
discord.py-slipstream-fork==2.6.0a0
numpy>=2.3,<3
opencv-python>=4.11.0.86,<5
Pillow>=9.2.0,<12
protobuf>=3.20.2,<7
python-dotenv>=1.1.0,<2
litellm@git+https://github.com/Slipstreamm/litellm@main
asyncpg>=0.29.0,<1.0
psycopg2-binary>=2.9.0,<3.0
redis
humanize
psutil
cryptography